requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
streamlit
sentence-transformers
python-dotenv
pandas